leetcode/0973.K_Closest_Points_to_Origin/K_Closest_Points_to_Origin_test.go
package K_Closest_Points_to_Origin

import (
	"reflect"
	"testing"
)

func Test_kClosest(t *testing.T) {
	type args struct {
		points [][]int
		K      int
	}
	tests := []struct {
		name string
		args args
		want [][]int
	}{
		// TODO: Add test cases.
		{
			"case",
			args{
				points: [][]int{{1, 3}, {-2, 2}, {2, -2}},
				K:      2,
			},
			[][]int{
				{-2, 2}, {2, -2},
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := kClosest(tt.args.points, tt.args.K); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("kClosest() = %v, want %v", got, tt.want)
			}
		})
	}
}
